extract methods to Commons lib
