+++ modules/engines/engine-hdb/pom.xml
@@ -17,11 +17,6 @@
   <packaging>jar</packaging>
 
   <dependencies>
-    <!--<dependency>
-      <groupId>com.sap.xsk</groupId>
-      <artifactId>xsk-modules-engines-core</artifactId>
-      <version>0.7.0-SNAPSHOT</version>
-    </dependency>-->
     <dependency>
       <groupId>com.sap.xsk</groupId>
       <artifactId>xsk-modules-engines-commons</artifactId>
